solved warnings
[Defect / Enhancement / New Feature]
Defect

[Problem]
warning: comparison between pointer and integer [enabled by default]

[Solution]
proper cast applied

[Reviewers]
vikram, naga

[Testing]
    [Test Cases]
    kernel booting
    suspend-resume

    [Label/CL# Synced To]
    sdb-common-android-jb-4.2

    [Platform]
    hawaiistone

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/13131
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoBuild Status <autobuild_status_noreply@broadcom.com>
Tested-by: AutoTest Status <autotest_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ drivers/crypto/brcm_spum_aes.c
@@ -333,9 +333,11 @@ static int spum_aes_dma_xfer(struct spum_aes_device *dd)
 	    PERIPHERAL_FLUSHP_END | DMA_PERI_REQ_ALWAYS_BURST;
 
 	rx_fifo =
-	    (u32)HW_IO_VIRT_TO_PHYS(dd->io_axi_base) + SPUM_AXI_FIFO_IN_OFFSET;
+		(u32)HW_IO_VIRT_TO_PHYS((u32)dd->io_axi_base) +
+		SPUM_AXI_FIFO_IN_OFFSET;
 	tx_fifo =
-	    (u32)HW_IO_VIRT_TO_PHYS(dd->io_axi_base) + SPUM_AXI_FIFO_OUT_OFFSET;
+		(u32)HW_IO_VIRT_TO_PHYS((u32)dd->io_axi_base) +
+		SPUM_AXI_FIFO_OUT_OFFSET;
 
 	INIT_LIST_HEAD(&head_in);
 	INIT_LIST_HEAD(&head_out);
